+++ packages/media-avatar-picker/stories/image-navigator/image-navigator-story.tsx
@@ -1,9 +1,11 @@
 /* tslint:disable:variable-name */
 import { storiesOf } from '@kadira/storybook';
 import * as React from 'react';
+import {Component} from 'react';
 import {tallImage, remoteImage} from '@atlaskit/media-test-helpers';
+import FieldText from '@atlaskit/field-text';
 import {ImageNavigator} from '../../src';
-import {CoverWrapper, Image} from './styled';
+import {CoverWrapper, Image, FieldTextWrapper} from './styled';
 
 let onLoadParams;
 let imageElement;
@@ -22,7 +24,47 @@ const handleImgRef = (img) => {
   imageElement = img;
 };
 
-const imageUrl = 'http://pic.templetons.com/brad/pano/midpano/center-wide.jpg';
+const remoteImages = [
+  'http://pic.templetons.com/brad/pano/midpano/center-wide.jpg',
+  'https://www.easycalculation.com/area/images/big-square.gif'
+];
+
+class CoverImage extends Component<{}, {}> {
+  state = {
+    imageSource: remoteImages[0]
+  };
+
+  onImageSourceChange = e => {
+    const imageSource = e.target.value;
+    this.setState({imageSource});
+  }
+
+  toggleImage = () => {
+    const {imageSource} = this.state;
+    const currentIndex = remoteImages.indexOf(imageSource);
+    const index = currentIndex >= remoteImages.length - 1 ? 0 : currentIndex + 1;
+
+    this.setState({imageSource: remoteImages[index]});
+  }
+
+  render() {
+    const {imageSource} = this.state;
+
+    return (
+      <div>
+        <FieldTextWrapper>
+          <FieldText value={imageSource} label="image source" onChange={this.onImageSourceChange} shouldFitContainer={true} />
+          <button onClick={this.toggleImage}>Toggle image</button>
+        </FieldTextWrapper>
+        <CoverWrapper>
+          <ImageNavigator allowZooming={false} mask="none" imageSource={imageSource} containerWidth="auto" onLoad={onLoad} />
+        </CoverWrapper>
+        <button onClick={exportImage}>Export</button>
+        <Image src="" alt="" innerRef={handleImgRef} />
+      </div>
+    );
+  }
+}
 
 storiesOf('Image navigator', {})
   .add('Local image', () => (
@@ -33,13 +75,7 @@ storiesOf('Image navigator', {})
     </div>
   ))
   .add('Auto width => cover image use case', () => (
-    <div>
-      <CoverWrapper>
-        <ImageNavigator allowZooming={false} mask="none" imageSource={imageUrl} containerWidth="auto" onLoad={onLoad} />
-      </CoverWrapper>
-      <button onClick={exportImage}>Export</button>
-      <Image src="" alt="" innerRef={handleImgRef} />
-    </div>
+    <CoverImage />
   ))
   .add('Remote image', () => (
     <div>